Simplify builders
