+++ moors/src/algorithms/mod.rs
@@ -1,15 +1,16 @@
+mod builder;
 pub(crate) mod helpers;
+mod macros;
 mod moo;
 mod soo;
 
+pub use builder::{AlgorithmBuilder, AlgorithmBuilderError, GeneticAlgorithm};
 pub use moo::agemoea::{AgeMoea, AgeMoeaBuilder};
 pub use moo::nsga2::{Nsga2, Nsga2Builder};
 pub use moo::nsga3::{Nsga3, Nsga3Builder};
 pub use moo::revea::{Revea, ReveaBuilder};
 pub use moo::rnsga2::{Rnsga2, Rnsga2Builder};
 pub use moo::spea2::{Spea2, Spea2Builder};
-pub use moo::{AlgorithmMOOBuilder, AlgorithmMOOBuilderError, GeneticAlgorithmMOO};
-pub use soo::{AlgorithmSOOBuilder, AlgorithmSOOBuilderError, GeneticAlgorithmSOO};
 
 pub(crate) use helpers::AlgorithmContext;
 pub use helpers::{AlgorithmError, InitializationError};
