+++ moors/src/lib.rs
@@ -103,10 +103,9 @@ mod private;
 pub mod random;
 
 pub use algorithms::{
-    AgeMoea, AgeMoeaBuilder, AlgorithmError, AlgorithmMOOBuilder, AlgorithmMOOBuilderError,
-    AlgorithmSOOBuilder, AlgorithmSOOBuilderError, GeneticAlgorithmMOO, GeneticAlgorithmSOO,
-    InitializationError, Nsga2, Nsga2Builder, Nsga3, Nsga3Builder, Revea, ReveaBuilder, Rnsga2,
-    Rnsga2Builder, Spea2, Spea2Builder,
+    AgeMoea, AgeMoeaBuilder, AlgorithmBuilder, AlgorithmBuilderError, AlgorithmError,
+    GeneticAlgorithm, InitializationError, Nsga2, Nsga2Builder, Nsga3, Nsga3Builder, Revea,
+    ReveaBuilder, Rnsga2, Rnsga2Builder, Spea2, Spea2Builder,
 };
 pub use duplicates::{
     CloseDuplicatesCleaner, ExactDuplicatesCleaner, NoDuplicatesCleaner, PopulationCleaner,
